fix: stats fixed problems

+++ stats-service/dto/src/main/java/ru/yandex/practicum/ewm/dto/ViewStatsDto.java
@@ -11,6 +11,6 @@ public class ViewStatsDto {
 
     private String uri;
 
-    private Integer hits;
+    private Long hits;
 
 }
